+++ changelogs/unreleased/6484-kaovilai
@@ -1,0 +1,1 @@
+Velero Plugins no longer need kopia indirect dependency in their go.mod
